backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.13.0
scikit-learn==1.6.1
numpy==2.2.0
python-dotenv==1.0.1
pydantic==2.10.6
httpx==0.28.1
